fix: resolve db, auth handling, and Celery issues

+++ tasks.py
@@ -8,16 +8,14 @@ load_dotenv()
 
 app=Celery('task',broker=os.getenv("CELERY_URL"))
 
-from database import Sessionlocal
-from Schema import Driver
 
 @app.task
 def del_user(driver_id):
     session = Sessionlocal()
     try:
-       element=session.query(Driver).filter(Driver.id==driver_id).one()
-       session.delete(element)
-       session.commit()
-       
+       element=session.query(Driver).filter(Driver.id==driver_id).one_or_none()
+       if element is None:
+        session.delete(element)
+        session.commit()
     finally:
         session.close()
